update gio hang + db

+++ src/controller/AddCtrl.java
@@ -34,7 +34,7 @@ public class AddCtrl extends HttpServlet {
 		response.setCharacterEncoding("utf-8");
 		String name = request.getParameter("nameb");
 		String author = request.getParameter("authorb");
-		float amount = Float.parseFloat(request.getParameter("amountb"));
+		int amount = Integer.parseInt(request.getParameter("amountb"));
 		int number = Integer.parseInt(request.getParameter("numberb"));
 		float weight = Float.parseFloat(request.getParameter("weightb"));
 		int year = Integer.parseInt(request.getParameter("yearb"));
